restructure for backwards compatibility
