+++ src/commands/token/flex.js
@@ -2,9 +2,9 @@ const { flags } = require('@oclif/command');
 const { TwilioClientCommand } = require('@twilio/cli-core').baseCommands;
 const { TwilioCliError } = require('@twilio/cli-core').services.error;
 const Twilio = require('twilio');
-const createToken = require('../../../helpers/accessToken.js');
-const globalFlags = require('../../../helpers/globalFlags.js');
-const { taskrouterFlags, validateWorkerSid, validateWorkspaceSid } = require('../../../helpers/taskrouterGlobals.js');
+const createToken = require('../../helpers/accessToken.js');
+const globalFlags = require('../../helpers/globalFlags.js');
+const { taskrouterFlags, validateWorkerSid, validateWorkspaceSid } = require('../../helpers/taskrouterGlobals.js');
 
 class FlexTokenGenerator extends TwilioClientCommand {
   constructor(argv, config, secureStorage) {
