+++ src/commands/token/voice.js
@@ -2,9 +2,9 @@ const { flags } = require('@oclif/command');
 const { TwilioClientCommand } = require('@twilio/cli-core').baseCommands;
 const { TwilioCliError } = require('@twilio/cli-core').services.error;
 const Twilio = require('twilio');
-const createToken = require('../../../helpers/accessToken.js');
-const globalFlags = require('../../../helpers/globalFlags.js');
-const { voiceFlags, validateTwimlAppSid } = require('../../../helpers/voiceGlobals.js');
+const createToken = require('../../helpers/accessToken.js');
+const globalFlags = require('../../helpers/globalFlags.js');
+const { voiceFlags, validateTwimlAppSid } = require('../../helpers/voiceGlobals.js');
 
 class VoiceTokenGenerator extends TwilioClientCommand {
   constructor(argv, config, secureStorage) {
